feat: add support for external docker networks
When having other docker compose files that submit jobs to the lava
server within this docker compose file, it's useful to be able to set
the external docker network.

This change introduces a new option that allows you to specify
external_docker_networks in a list and for each network name you add to
your config file, your services will be connected to that docker
network.